+++ DatadogSessionReplay/Tests/SessionReplayTests.swift
@@ -64,6 +64,7 @@ class SessionReplayTests: XCTestCase {
         let sr = try XCTUnwrap(core.get(feature: SessionReplayFeature.self))
         XCTAssertEqual(sr.recordingCoordinator.sampler.samplingRate, 42)
         XCTAssertEqual(sr.recordingCoordinator.privacy, .mask)
+        XCTAssertEqual(sr.recordingCoordinator.imagePrivacy, .maskAll)
         XCTAssertEqual(sr.recordingCoordinator.touchPrivacy, .hide)
         XCTAssertNil((sr.requestBuilder as? SegmentRequestBuilder)?.customUploadURL)
         let r = try XCTUnwrap(core.get(feature: ResourcesFeature.self))
@@ -72,7 +73,7 @@ class SessionReplayTests: XCTestCase {
 
     func testWhenEnabledWithDefaultConfigurationWithNewAPI() throws {
         let textAndInputPrivacy: SessionReplayTextAndInputPrivacyLevel = .mockRandom()
-        let touchPrivacy: SessionReplayTouchPrivacyLevel = .mockRandom()
+        let touchPrivacy: TouchPrivacyLevel = .mockRandom()
         config = SessionReplay.Configuration(replaySampleRate: 42, textAndInputPrivacyLevel: textAndInputPrivacy, touchPrivacyLevel: touchPrivacy)
 
         // When
@@ -133,7 +134,7 @@ class SessionReplayTests: XCTestCase {
         config.defaultPrivacyLevel = randomPrivacy
         let textAndInputPrivacy: SessionReplayTextAndInputPrivacyLevel = .mockRandom()
         config.textAndInputPrivacyLevel = textAndInputPrivacy
-        let randomTouchPrivacy: SessionReplayTouchPrivacyLevel = .mockRandom()
+        let randomTouchPrivacy: TouchPrivacyLevel = .mockRandom()
         config.touchPrivacyLevel = randomTouchPrivacy
 
         // When
@@ -153,7 +154,7 @@ class SessionReplayTests: XCTestCase {
         config.defaultPrivacyLevel = randomPrivacy
         let randomTextAndInputPrivacy: SessionReplayTextAndInputPrivacyLevel = .mockRandom()
         config.textAndInputPrivacyLevel = randomTextAndInputPrivacy
-        let randomTouchPrivacy: SessionReplayTouchPrivacyLevel = .mockRandom()
+        let randomTouchPrivacy: TouchPrivacyLevel = .mockRandom()
         config.touchPrivacyLevel = randomTouchPrivacy
 
         // When